fix: too early running of FirstWorldInit bucket
Running it during `world_state.onInitUI` causes the bucket to run before `world_state.init()` has been called. This causes issues with many objects instantiated during this bucket especially with rosters/players. The correct timing for the bucket is after the world_state has been fully initialized which means it should be run with a hook on `onInit` or `init` of `world_state`.