feat: call FirstWorldInit at the start of world_state.init
Calling it at the end of `onInit` is too late for the intended use case of this bucket because if things in this bucket set some global variables which are meant to be used during the loading/starting of a campaign then those things will not be ready in time. The reason this bucket exists is because it depends on certain global objects being instantiated, and those are available at the beginning of `init` without exception. Therefore, calling it at the beginning of `init` is safer and offers more flexibility.